requirements.txt
streamlit
ollama
chromadb
pandas
sentence-transformers
PyPDF2
python-docx
crewai
langchain
tqdm
pycryptodome